feat : user,auth 완료 user , auth 관련 api 를 완성했습니다. ( 테스트 전 )
